# upload/extracted/Attached-Assets-main/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Цифровой обходной лист — Digital Routing Sheet HR system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers:
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create and send an offer to a candidate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Offer created and sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferCreatedResponse"
  /offers/by-token/{token}:
    get:
      operationId: getOfferByToken
      tags: [offers]
      summary: Get offer info by candidate token (no auth required)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Offer info for candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferPublic"
        "404":
          description: Token not found or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /offers/{id}/accept:
    post:
      operationId: acceptOffer
      tags: [offers]
      summary: Candidate clicks "Accept offer" — triggers OTP send
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptOfferInput"
      responses:
        "200":
          description: OTP sent to candidate email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpSentResponse"
  /offers/{id}/verify-otp:
    post:
      operationId: verifyOfferOtp
      tags: [offers]
      summary: Candidate verifies OTP to confirm offer acceptance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: Offer accepted, routing sheet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferAcceptedResponse"
        "400":
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /offers/{id}/resend-token:
    post:
      operationId: resendOfferToken
      tags: [offers]
      summary: Re-issue the candidate access token
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: New token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"

  # ── Routing Sheets ────────────────────────────────────
...
  /email-templates:
    get:
      operationId: listEmailTemplates
      tags: [admin]
      summary: List all email templates (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Email templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailTemplate"
  /email-templates/{type}:
    get:
      operationId: getEmailTemplate
      tags: [admin]
      summary: Get a single email template (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Email template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEmailTemplate
      tags: [admin]
      summary: Update an email template's subject and/or body (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailTemplateUpdate"
      responses:
        "200":
          description: Updated template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailTemplate"

  # ── Audit Log (admin) ─────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── Auth ─────────────────────────────────────────────
...
    Offer:
      type: object
      required: [id, candidateId, sentById, status, createdAt]
      properties:
        id:
          type: number
        candidateId:
          type: number
        sentById:
          type: number
        status:
          type: string
          enum: [sent, otp_pending, accepted, expired]
        token:
          type: ["string", "null"]
        tokenExpiresAt:
          type: ["string", "null"]
          format: date-time
        acceptedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    OfferInput:
      type: object
      required: [candidateId, branchId, positionId]
      properties:
        candidateId:
          type: number
        branchId:
          type: number
        positionId:
          type: number
        message:
          type: string
    OfferPublic:
      type: object
      required: [id, candidateName, companyName, status]
      properties:
        id:
          type: number
        candidateName:
          type: string
        companyName:
          type: string
        status:
          type: string
          enum: [sent, otp_pending, accepted, expired]
        message:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
          format: date-time
    AcceptOfferInput:
      type: object
      required: [token, consentGiven]
      properties:
        token:
          type: string
        consentGiven:
          type: boolean
    VerifyOtpInput:
      type: object
      required: [token, otp]
      properties:
        token:
          type: string
        otp:
          type: string
    OfferCreatedResponse:
      type: object
      required: [id, candidateId, status, token, statusToken, routingSheetId, createdAt]
      properties:
        id:
          type: number
        candidateId:
          type: number
        status:
          type: string
          enum: [sent, otp_pending, accepted, expired]
        token:
          type: string
        statusToken:
          type: string
        routingSheetId:
          type: number
        createdAt:
          type: string
          format: date-time
    OtpSentResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    OfferAcceptedResponse:
      type: object
      required: [message, routingSheetId, statusToken]
      properties:
        message:
          type: string
        routingSheetId:
          type: number
        statusToken:
          type: string

    # ── Routing Sheets ────────────────────────────────────
...
    EmailTemplate:
      type: object
      required: [id, templateType, subject, bodyHtml, updatedAt]
      properties:
        id:
          type: number
        templateType:
          type: string
        subject:
          type: string
        bodyHtml:
          type: string
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: ["string", "null"]
        variables:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
    EmailTemplateUpdate:
      type: object
      properties:
        subject:
          type: string
        bodyHtml:
          type: string
...